implemented a transaction into the supplier's account

+++ src/main/java/com/example/exam9/service/ProviderService.java
@@ -9,6 +9,6 @@ import java.util.List;
 public interface ProviderService {
 
     List<ProviderDto> getProviders();
-    ProviderDto getProviderById(Long id);
+    ProviderDto getProviderById(Integer id);
 
 }
