+++ src/main/java/com/example/exam9/service/ProviderUsersService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 @Service
 public interface ProviderUsersService {
 
-    Integer createUniqueAccount(Integer userId, Long providerId);
+    Integer createUniqueAccount(Integer userId, Integer providerId);
 
-    Integer getUniqueNumber(Integer personalAccountNumber, Long providerId);
+    Integer getUniqueNumber(Integer personalAccountNumber, Integer providerId);
 }
